Dtos For Order iS Completed

+++ src/main/java/com/fullstackwebapp/DTO/OrderItemRequest.java
@@ -3,7 +3,7 @@ package com.fullstackwebapp.DTO;
 import com.fullstackwebapp.Model.Product;
 
 public record OrderItemRequest(
-        int ProductId,
+        Long ProductId,
         int Quantity
 ) {
 }
